1.更新版本号到1.2.0 2.优化代码结构 3.添加32位riscv架构的支持,支持spike模拟器

+++ apps/syscommand/syscommand.c
@@ -89,7 +89,7 @@ void shell_lscpu(long argc, char *argv[])
         }
         else
         {
-            printf("frequency:   %ld Hz\n", infos[i].frequency);
+            printf("frequency:   %lld Hz\n", (long long)infos[i].frequency);
         }
         printf("\n");
     }
